+++ packages/terra-list/src/List.jsx
@@ -85,6 +85,11 @@ const propTypes = {
    */
   isDraggable: PropTypes.bool,
   /**
+   * ![IMPORTANT](https://badgen.net/badge/UX/Accessibility/blue)
+   * Whether or not the list item is focusable with Tab key. Ensure alternative way of focusing list item when set to true for best accessibility experience.
+   */
+  isTabFocusDisabled: PropTypes.bool,
+  /**
    * Function callback when the Item is dropped. Parameters:
    * @param {Object} result result
    * @param {Object} provided provided
@@ -93,11 +98,12 @@ const propTypes = {
 };
 
 const defaultProps = {
+  ariaSelectionStyle: 'none',
   children: [],
   dividerStyle: 'none',
+  isTabFocusDisabled: false,
   paddingStyle: 'none',
   role: 'none',
-  ariaSelectionStyle: 'none',
 };
 
 const List = ({
@@ -112,6 +118,7 @@ const List = ({
   role,
   ariaSelectionStyle,
   isDraggable,
+  isTabFocusDisabled,
   onDragEnd,
   ...customProps
 }) => {
@@ -238,10 +245,15 @@ const List = ({
 
   const cloneListItem = (ListItem, provider) => React.cloneElement(ListItem, {
     isDraggable: ListItem?.props?.isSelectable,
+    isTabFocusDisabled,
     refCallback: provider.innerRef,
     ...provider.draggableProps,
     ...provider.dragHandleProps,
   });
+
+  const clone = (object) => React.Children.map(object, (listitem) => React.cloneElement(listitem, {
+    isTabFocusDisabled,
+  }));
 
   const renderListDom = () => (
     // eslint-disable-next-line jsx-a11y/no-noninteractive-element-interactions, jsx-a11y/role-supports-aria-props
@@ -249,13 +261,14 @@ const List = ({
       {...customProps}
       {...attrSpread}
       aria-describedby={ariaDescribedBy}
-      aria-description={ariaDescription} // eslint-disable-line jsx-a11y/aria-props
+       // eslint-disable-next-line jsx-a11y/aria-props
+      aria-description={ariaDescription}
       aria-details={ariaDetails}
       className={listClassNames}
       ref={handleListRef}
       onKeyDown={handleKeyDown}
     >
-      {children}
+      {clone(children)}
     </ul>
   );
 
@@ -268,13 +281,14 @@ const List = ({
         )}
       >
         {(provided) => (
-          // eslint-disable-next-line jsx-a11y/no-noninteractive-element-interactions, jsx-a11y/role-supports-aria-props
+          /* eslint-disable-next-line jsx-a11y/no-noninteractive-element-interactions, jsx-a11y/role-supports-aria-props */
           <ul
             {...provided.droppableProps}
             {...customProps}
             {...attrSpread}
             aria-describedby={ariaDescribedBy}
-            aria-description={ariaDescription} // eslint-disable-line jsx-a11y/aria-props
+            // eslint-disable-next-line jsx-a11y/aria-props
+            aria-description={ariaDescription}
             aria-details={ariaDetails}
             className={listClassNames}
             ref={(refobj) => {