+++ packages/terra-dropdown-button/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+## 1.37.0 - (October 26, 2023)
+
+* Changed
+  * Updated focus border as dashed border in `dropdown-split-button`.
+
 ## 1.36.0 - (August 8, 2023)
 
 * Added